+++ Runtime/Scripts/VisualScripting/AnalyticSendDataUnit.cs
@@ -1,5 +1,5 @@
-using Reflectis.SDK.Core.SystemFramework;
-using Reflectis.SDK.ReflectisApi;
+using Virtuademy.SDK.Core.SystemFramework;
+using Virtuademy.SDK.PlatformApi;
 
 using System;
 using System.Collections.Generic;
@@ -10,7 +10,7 @@ using Unity.VisualScripting;
 
 using UnityEngine;
 
-namespace Reflectis.CreatorKit.Worlds.Analytics
+namespace Virtuademy.CreatorKit.Worlds.Analytics
 {
     [UnitTitle(UNIT_TITLE)]
     [UnitSurtitle("Reflectis Analytic")]