+++ Runtime/Scripts/VisualScripting/DisplayableData/DisplayableContentBase.cs
@@ -1,8 +1,8 @@
-using Reflectis.SDK.Core.Utilities;
+using Virtuademy.SDK.Core.Utilities;
 
 using System.Collections.Generic;
 
-namespace Reflectis.CreatorKit.Worlds.Analytics
+namespace Virtuademy.CreatorKit.Worlds.Analytics
 {
     public abstract class DisplayableContentBase
     {